refactor(send): unify AutoSizeInput cross-platform props
